feat: add RegisterForm component and update RegisterPage for admin registration
